Let the ot. as the start point for scan,

+++ include/sac_proc.h
@@ -4,4 +4,4 @@
 #pragma once
 
 /* */
-int sac_proc_station_data_extract( const char *, const char *, const float [], float *[], int *, float *, float * );
+int sac_proc_station_data_extract( const char *, const char *, const float [], float *[], int *, float *, double * );
